Extracted API and Test modules
This fixes #55
