+++ logbook-api/src/main/java/org/zalando/logbook/HttpLogWriter.java
