+++ logbook-api/src/main/java/org/zalando/logbook/Obfuscator.java
