+++ logbook-core/src/main/resources/META-INF/services/org.zalando.logbook.LogbookFactory
@@ -1,0 +1,1 @@
+org.zalando.logbook.DefaultLogbookFactory
