+++ logbook-spring-boot-starter/pom.xml
@@ -58,6 +58,10 @@
 
         <!-- testing -->
         <dependency>
+            <groupId>org.zalando</groupId>
+            <artifactId>logbook-test</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-simple</artifactId>
         </dependency>
