+++ logbook-test/pom.xml
@@ -13,9 +13,9 @@
         <version>0.15.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>logbook-core</artifactId>
+    <artifactId>logbook-test</artifactId>
 
-    <name>Logbook: Core</name>
+    <name>Logbook: Test</name>
     <description>HTTP request and response logging</description>
 
     <scm>
@@ -24,33 +24,14 @@
         <developerConnection>scm:git:git@github.com:zalando//logbook.git</developerConnection>
     </scm>
 
+    <properties>
+        <maven.deploy.skip>true</maven.deploy.skip>
+    </properties>
+
     <dependencies>
         <dependency>
             <groupId>org.zalando</groupId>
             <artifactId>logbook-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <optional>true</optional>
-        </dependency>
-
-        <!-- testing -->
-        <dependency>
-            <groupId>org.zalando</groupId>
-            <artifactId>logbook-test</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-nop</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.github.stefanbirkner</groupId>
-            <artifactId>system-rules</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.jayway.jsonpath</groupId>
-            <artifactId>json-path-assert</artifactId>
         </dependency>
     </dependencies>
