+++ pom.xml
@@ -52,10 +52,12 @@
     </developers>
 
     <modules>
+        <module>logbook-api</module>
         <module>logbook-core</module>
         <module>logbook-httpclient</module>
         <module>logbook-servlet</module>
         <module>logbook-spring-boot-starter</module>
+        <module>logbook-test</module>
     </modules>
 
     <properties>
@@ -77,6 +79,11 @@
             <!-- logbook -->
             <dependency>
                 <groupId>org.zalando</groupId>
+                <artifactId>logbook-api</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.zalando</groupId>
                 <artifactId>logbook-core</artifactId>
                 <version>${project.version}</version>
             </dependency>
@@ -94,6 +101,12 @@
                 <groupId>org.zalando</groupId>
                 <artifactId>logbook-spring-boot-starter</artifactId>
                 <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.zalando</groupId>
+                <artifactId>logbook-test</artifactId>
+                <version>${project.version}</version>
+                <scope>test</scope>
             </dependency>
 
             <!-- common -->
